完善检查组的新建和修改功能

+++ health-reservation/src/main/java/com/health/reservation/service/impl/TCheckgroupServiceImpl.java
@@ -2,11 +2,17 @@ package com.health.reservation.service.impl;
 
 import java.util.List;
 import com.health.common.utils.DateUtils;
+import com.health.common.utils.bean.BeanUtils;
+import com.health.reservation.dto.TCheckgroupDto;
+import com.health.reservation.mapper.TCheckgroupCheckitemMapper;
+import com.health.reservation.mapper.TSetmealCheckgroupMapper;
+import com.health.reservation.vo.TCheckgroupVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.health.reservation.mapper.TCheckgroupMapper;
 import com.health.reservation.domain.TCheckgroup;
 import com.health.reservation.service.ITCheckgroupService;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * 检查组管理Service业务层处理
@@ -14,11 +20,16 @@ import com.health.reservation.service.ITCheckgroupService;
  * @author 1926
  * @date 2025-11-08
  */
+@Transactional
 @Service
-public class TCheckgroupServiceImpl implements ITCheckgroupService 
+public  class TCheckgroupServiceImpl implements ITCheckgroupService
 {
     @Autowired
     private TCheckgroupMapper tCheckgroupMapper;
+    @Autowired
+    private TCheckgroupCheckitemMapper tCheckgroupCheckitemMapper;
+    @Autowired
+    private TSetmealCheckgroupMapper tSetmealCheckgroupMapper;
 
     /**
      * 查询检查组管理
@@ -27,9 +38,16 @@ public class TCheckgroupServiceImpl implements ITCheckgroupService
      * @return 检查组管理
      */
     @Override
-    public TCheckgroup selectTCheckgroupById(Long id)
-    {
-        return tCheckgroupMapper.selectTCheckgroupById(id);
+    public TCheckgroupVo selectTCheckgroupById(Long id)
+    { //1.查询检查组基本信息
+        TCheckgroup tCheckgroup = tCheckgroupMapper.selectTCheckgroupById(id);
+        //2.根据检查组的id查询检查项关联数据
+        List<Long> list=tCheckgroupCheckitemMapper.selectCheckitemIds(id);
+        //3.封装vo对象
+        TCheckgroupVo vo=new TCheckgroupVo();
+        BeanUtils.copyProperties(tCheckgroup,vo);
+        vo.setCheckItemIds(list);
+        return vo;
     }
 
     /**
@@ -44,30 +62,54 @@ public class TCheckgroupServiceImpl implements ITCheckgroupService
         return tCheckgroupMapper.selectTCheckgroupList(tCheckgroup);
     }
 
+    @Override
+    public int insertTCheckgroup(TCheckgroup dto) {
+        return 0;
+    }
+
+
     /**
-     * 新增检查组管理
-     * 
-     * @param tCheckgroup 检查组管理
+     * 新增检查组
+     *
+     * @param dto 检查组
      * @return 结果
      */
+    @Transactional
     @Override
-    public int insertTCheckgroup(TCheckgroup tCheckgroup)
+    public int insertTCheckgroup(TCheckgroupDto dto)
     {
-        tCheckgroup.setCreateTime(DateUtils.getNowDate());
-        return tCheckgroupMapper.insertTCheckgroup(tCheckgroup);
+        dto.setCreateTime(DateUtils.getNowDate());
+        int count=tCheckgroupMapper.insertTCheckgroup(dto);
+        if (dto.getCheckItemIds()!=null&&dto.getCheckItemIds().size()>0){
+            tCheckgroupCheckitemMapper .insertCheckitemIds(dto.getCheckItemIds(),dto.getId());
+        }
+
+        return count;
     }
 
     /**
      * 修改检查组管理
      * 
-     * @param tCheckgroup 检查组管理
+     * @param dto 检查组管理
      * @return 结果
      */
     @Override
-    public int updateTCheckgroup(TCheckgroup tCheckgroup)
-    {
-        tCheckgroup.setUpdateTime(DateUtils.getNowDate());
-        return tCheckgroupMapper.updateTCheckgroup(tCheckgroup);
+    public int updateTCheckgroup(TCheckgroupDto dto) {
+        try {
+            //1.修改检查组基本信息
+            dto.setUpdateTime(DateUtils.getNowDate());
+            int count = tCheckgroupMapper.updateTCheckgroup(dto);
+            //2.修改检查组和检查项的关联关系
+            if (dto.getCheckItemIds() != null && dto.getCheckItemIds().size() > 0) {
+                //1.删除关联关系
+                tCheckgroupCheckitemMapper.deleteTCheckgroupCheckitemByCheckgroupId(dto.getId());
+                //2.保存关联关系
+                tCheckgroupCheckitemMapper.insertCheckitemIds(dto.getCheckItemIds(), dto.getId());
+            }
+            return count;
+        } catch (Exception e) {
+            throw new RuntimeException("修改检查组信息失败");
+        }
     }
 
     /**
@@ -90,7 +132,14 @@ public class TCheckgroupServiceImpl implements ITCheckgroupService
      */
     @Override
     public int deleteTCheckgroupById(Long id)
-    {
+    { //1。判读检查组是否被套餐关联
+        int count = tSetmealCheckgroupMapper.selectCountByCheckgroupId(id);
+        if (count > 0){
+            throw new RuntimeException("当前检查组被套餐关联，不能删除");
+        }
+        //2.删除检查组和检查项的关联关系
+        tCheckgroupCheckitemMapper.deleteTCheckgroupCheckitemByCheckgroupId(id);
+        //3.删除检查组信息
         return tCheckgroupMapper.deleteTCheckgroupById(id);
     }
 }
